agent/requirements.txt
# TaskFlow Customer Support Agent
google-adk>=0.3.0
python-dotenv>=1.0.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
litellm>=1.50.0
